# openapi.json
{"openapi":"3.1.0","info":{"title":"Ghigus API","version":"0.1"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/cases/{slug}":{"delete":{"tags":["cases"],"summary":"Delete Case (and related data, NOT maps)","description":"Cancella la pratica 'slug' e tutte le righe collegate (uploads, riclass, bilancio contabile, kpi, ecc.),\nSENZA toccare le mappe (stg_map_*). Ogni tabella viene pulita in una transazione separata.","operationId":"delete_case_cases__slug__delete","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string","title":"Slug"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["cases"],"summary":"Get Case","operationId":"get_case_cases__slug__get","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string","title":"Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{slug}/upload-bc":{"post":{"tags":["uploads"],"summary":"Upload Bc","operationId":"upload_bc_cases__slug__upload_bc_post","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string","title":"Slug"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_bc_cases__slug__upload_bc_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{slug}/upload-tb":{"post":{"tags":["uploads"],"summary":"Upload Tb","operationId":"upload_tb_cases__slug__upload_tb_post","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string","title":"Slug"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_tb_cases__slug__upload_tb_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{slug}/upload-xbrl":{"post":{"tags":["uploads"],"summary":"Upload Xbrl","operationId":"upload_xbrl_cases__slug__upload_xbrl_post","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string","title":"Slug"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_xbrl_cases__slug__upload_xbrl_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases":{"get":{"tags":["cases"],"summary":"List Cases","operationId":"list_cases_cases_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CaseOut"},"type":"array","title":"Response List Cases Cases Get"}}}}}},"post":{"tags":["cases"],"summary":"Create Case","operationId":"create_case_cases_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CaseOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{slug}/sp-riclass":{"get":{"tags":["riclass"],"summary":"List Sp Riclass","operationId":"list_sp_riclass_cases__slug__sp_riclass_get","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string","title":"Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SpRiclassOut"},"title":"Response List Sp Riclass Cases  Slug  Sp Riclass Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{slug}/ce-riclass":{"get":{"tags":["riclass"],"summary":"List Ce Riclass","operationId":"list_ce_riclass_cases__slug__ce_riclass_get","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string","title":"Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CeRiclassOut"},"title":"Response List Ce Riclass Cases  Slug  Ce Riclass Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{slug}/kpi-standard":{"get":{"tags":["riclass"],"summary":"List Kpi","operationId":"list_kpi_cases__slug__kpi_standard_get","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string","title":"Slug"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/KpiStandardOut"},"title":"Response List Kpi Cases  Slug  Kpi Standard Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/maps/import-csv":{"post":{"tags":["maps"],"summary":"Import Maps Csv Global","operationId":"import_maps_csv_global_maps_import_csv_post","parameters":[{"name":"replace","in":"query","required":false,"schema":{"type":"boolean","description":"Se True, cancella le mappe esistenti nello scope scelto.","default":true,"title":"Replace"},"description":"Se True, cancella le mappe esistenti nello scope scelto."},{"name":"slug","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Se valorizzato, importa per quello slug; altrimenti globale.","title":"Slug"},"description":"Se valorizzato, importa per quello slug; altrimenti globale."}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_import_maps_csv_global_maps_import_csv_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cases/{slug}/maps/import-csv":{"post":{"tags":["maps"],"summary":"Import Maps Csv Case","operationId":"import_maps_csv_case_cases__slug__maps_import_csv_post","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string","title":"Slug"}},{"name":"replace","in":"query","required":false,"schema":{"type":"boolean","description":"Se True, cancella le mappe esistenti del case prima di importare.","default":true,"title":"Replace"},"description":"Se True, cancella le mappe esistenti del case prima di importare."}],"requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_import_maps_csv_case_cases__slug__maps_import_csv_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_import_maps_csv_case_cases__slug__maps_import_csv_post":{"properties":{"file_sp":{"anyOf":[{"type":"string","format":"binary"},{"type":"null"}],"title":"File Sp","description":"CSV SP"},"file_ce":{"anyOf":[{"type":"string","format":"binary"},{"type":"null"}],"title":"File Ce","description":"CSV CE"}},"type":"object","title":"Body_import_maps_csv_case_cases__slug__maps_import_csv_post"},"Body_import_maps_csv_global_maps_import_csv_post":{"properties":{"file_sp":{"anyOf":[{"type":"string","format":"binary"},{"type":"null"}],"title":"File Sp","description":"CSV SP: imported_code,imported_desc,riclass_code,riclass_desc"},"file_ce":{"anyOf":[{"type":"string","format":"binary"},{"type":"null"}],"title":"File Ce","description":"CSV CE: imported_code,imported_desc,riclass_code,riclass_desc"}},"type":"object","title":"Body_import_maps_csv_global_maps_import_csv_post"},"Body_upload_bc_cases__slug__upload_bc_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_bc_cases__slug__upload_bc_post"},"Body_upload_tb_cases__slug__upload_tb_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_tb_cases__slug__upload_tb_post"},"Body_upload_xbrl_cases__slug__upload_xbrl_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_xbrl_cases__slug__upload_xbrl_post"},"CaseCreate":{"properties":{"slug":{"type":"string","title":"Slug"},"name":{"type":"string","title":"Name"},"company_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company Id"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["slug","name"],"title":"CaseCreate"},"CaseOut":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"company_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company Id"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"slug":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Slug"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},"type":"object","required":["id","name"],"title":"CaseOut"},"CeRiclassOut":{"properties":{"id":{"type":"integer","title":"Id"},"case_id":{"type":"string","title":"Case Id"},"riclass_code":{"type":"string","title":"Riclass Code"},"riclass_desc":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Riclass Desc"},"amount":{"type":"number","title":"Amount"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","case_id","riclass_code","amount","created_at"],"title":"CeRiclassOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"KpiStandardOut":{"properties":{"id":{"type":"integer","title":"Id"},"case_id":{"type":"string","title":"Case Id"},"name":{"type":"string","title":"Name"},"value":{"type":"number","title":"Value"},"unit":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Unit"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","case_id","name","value","created_at"],"title":"KpiStandardOut"},"SpRiclassOut":{"properties":{"id":{"type":"integer","title":"Id"},"case_id":{"type":"string","title":"Case Id"},"riclass_code":{"type":"string","title":"Riclass Code"},"riclass_desc":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Riclass Desc"},"amount":{"type":"number","title":"Amount"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","case_id","riclass_code","amount","created_at"],"title":"SpRiclassOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}